requirements.txt
fpdf2~=2.8.2